Registration page, some permissions management and registration confirmation to new users.

+++ server.js
@@ -11,7 +11,7 @@ require("dotenv").config({ path: ".env" });
 
  const accountSid = process.env.TWILIO_ACCOUNT_SID;
  const AuthToken = process.env.TWILIO_AUTH_TOKEN;
- const receiverNumber = process.env.bossnumber;
+ const receiverNumber = process.env.mynumber;
  const senderNumber = process.env.twilioNumber;
 
 // const client = require('twilio')(accountSid, AuthToken);
@@ -46,9 +46,9 @@ app.use(express.static("public"));
     
 app.use(clientSessions({
     cookieName: "session",
-    secret: "web322_week10_demoSession",
-    duration: 20*60*1000,
-    activeDuration: 1000*60
+    secret: "Patties@Express_Supper!Reporter#Cookie)Session@Webpager...",
+    duration: 60*60*1000,//1 hour session
+    activeDuration: 1000*60*30
 }));
 
 app.use(bodyParser.urlencoded({extended: false }));
@@ -77,6 +77,7 @@ function ensureLogin(req, res, next) {
       next();
     }
 };
+//https://stormpath.com/blog/everything-you-ever-wanted-to-know-about-node-dot-js-sessions
 /* #endregion */
 
 /* #region ROUTES */
@@ -94,7 +95,7 @@ app.post("/login", (req, res) => {
     const username = req.body.username;
     const password = req.body.password;
 
-    if (username === "" || password === "") {
+    if (username === "" || password === "" || username === "^*;" || password === "^*;" ) {
         return res.render("login", {errorMsg: "Missing Credentials.", layout: false});
     }
 
@@ -105,7 +106,18 @@ app.post("/login", (req, res) => {
                 res.render("login", {errorMsg: "login does not exist!", layout: false});
             } else {
                 // user exists
-                if (username === usr.username && password === usr.password){
+                if (username === usr.username && password === usr.password && usr.isAdmin){
+                    req.session.user = {
+                        username: usr.username,
+                        email: usr.email,
+                        phone: usr.phone,
+                        firstName: usr.firstName,
+                        lastName: usr.lastName,
+                        isAdmin: usr.isAdmin
+                    };
+                    res.redirect("/report");
+                }
+                else if (username === usr.username && password === usr.password){
                     req.session.user = {
                         username: usr.username,
                         email: usr.email,
@@ -138,6 +150,175 @@ app.get("/Profile", ensureLogin, (req,res)=>{
 app.get("/Profile/Edit", ensureLogin, (req,res)=>{
     res.render("ProfileEdit", {user: req.session.user, layout: false});
 });
+
+app.get("/addProfile",/* ensureLogin,*/ (req,res)=>{
+    //if(req.session.user.isAdmin === true){
+        res.render("addProfile", {/*user: req.session.user, */layout: false});
+    //} else {
+    //    res.redirect("/Profile", session.persist);
+   // }
+});
+//https://stormpath.com/blog/everything-you-ever-wanted-to-know-about-node-dot-js-sessions
+
+
+app.get("/firstrunsetup", (req,res)=> {
+    var Yuri = new UserModel({
+        username: 'yuri',
+        password: 'password',
+        firstName: 'Yuri',
+        lastName: 'Yuri',
+        email: 'yhofalcao@gmail.com',
+        phone: '',
+        isAdmin: true
+    });
+    console.log("got here!");
+    Yuri.save((err)=> {
+        console.log("Error: " + err + ';');
+        if (err) {
+            console.log("There was an error creating Yuri: " + err);
+        } else {
+            console.log("Yuri was created");
+        }
+    });
+    console.log("got here 2!");
+    res.redirect("/");
+})
+
+
+/*app.post("/addProfile", function (req,res) {
+    const Form_data = req.body;
+    const addUser = {
+        username: Form_data.username, 
+        password: Form_data.password, 
+        firstName: Form_data.firstname,    
+        lastName: Form_data.lastname,          
+        Email: Form_data.email,          
+        Phone: Form_data.phone,        
+        isAdmin: false
+    };
+    const newUser = new UserModel(addUser);
+    newUser.save()
+})*/
+
+app.post("/AddProfile", function (req,res) {
+    //if(req.session.user.isAdmin === true){
+    var newUser = new UserModel({
+        username: req.body.username, 
+        password: req.body.password, 
+        firstName: req.body.firstname,    
+        lastName: req.body.lastname,          
+        email: req.body.email,          
+        phone: req.body.phone,        
+        isAdmin: false
+    });  
+    console.log("got here!");
+    newUser.save((err)=> {
+        console.log("Error: " + err + ';');
+        if (err) {
+            console.log("There was an error creating newUser: " + err);
+        } else {
+            console.log("newUser was created");
+        }
+    });
+    console.log("got here 2!");
+    res.redirect("/");
+
+        //user.save((err)=>{});
+
+    
+    var mailNewUser = {
+        from: process.env.mailerUser,
+        to: req.body.email,
+        subject: 'Welcome to Patties Reporting System',
+        html: 
+        "<p>Welcome to Patties Express Reporting System</p><br>"+
+        "<p>Please keep your profile information updated</p>" +
+        "<p>For security reasons, please update your password every semester</p> </br>" +
+        "<p>Since every report sent generates a SMS and e-mail message, please keep it minimal</p>" +
+        "<p>Do not hesitate to inquire about any doubts or issues. This service is in development, so please be patinet.</p>" +
+        "</br></br><p>By subscribing to this service you agree to share your profile data and reports data with Patties Express staff and with the general public.</p>" +
+        "<p>We make our best efforts to keep your data safe and private, but we cannot prevent all threats and data theft can happen.</p>"
+        
+    }
+
+    var smsMessage = {
+        body: "Welcome to Patties Reporting System" + 
+        "\nFor more information about the reporting system, please check your email (check also the spam) or visit our reporting system page at https://evening-savannah-18144.herokuapp.com/" 
+    }
+
+    client.messages.create({
+          to: req.body.phone, //to: '+12223333444'
+          from: senderNumber,
+          body: smsMessage.body
+    })
+
+    transporter.sendMail(mailNewUser, (error, info) => {
+        if (error){
+            console.log("ERROR: " + error);
+        } else {
+            console.log("SUCCESS: " + info.response);
+        }
+    });
+
+   // res.render("/ProfilesDashboard");
+//} 
+//else {
+ //   res.redirect("/Profile");
+//}
+});
+
+app.get("/ProfilesDashboard", ensureLogin, (req,res)=>{
+    //const user = req.session.user;
+    //const isAdmin = user.isAdmin;
+    if(req.session.user.isAdmin === true){
+        UserModel.find()
+        .lean()
+        .exec()
+        .then((users) =>{
+            res.render("ProfilesDashboard", {users: users, hasUser: !!users.length, user: req.session.user, layout: false});
+        });
+        //res.render("ProfilesDashboard", {user: req.session.user, layout: false});
+    } else {
+        res.redirect("/Profile");
+    }
+});
+
+app.get("/report", ensureLogin, (req,res) => {
+    ReportModel.find()
+        .lean()
+        .exec()
+        .then((report) =>{
+            res.render("report", {report: report, hasReport: !!report.length, user: req.session.user, layout: false});
+        });
+})
+
+app.get("/Profile/Edit/:username", ensureLogin, (req,res) => {
+    const username = req.params.username;
+    if(req.session.user.isAdmin === true){
+        UserModel.findOne({username: username})
+            .lean()
+            .exec()
+            .then((user)=>{
+                res.render("ProfileEdit", {user: req.session.user, user: user, editmode: true, layout: false})
+            .catch((err)=>{});
+        });
+    } else {
+        res.redirect("/ProfileDashboard", {user: req.session.user});
+    }
+});
+
+app.get("/Profile/Delete/:username", ensureLogin, (req, res) => {
+    if(req.session.user.isAdmin === true){
+    const username = req.params.username;
+    ReportModel.deleteOne({username: username})
+        .then(()=>{
+            res.redirect("/ProfilesDashboard");
+        });
+    } else {
+        res.redirect("/Profile");
+    }
+})
+
 
 app.post("/Profile/Edit", ensureLogin, (req,res) => {
     const username = req.body.username;
@@ -289,11 +470,15 @@ app.get("/report/Edit/:reportID", ensureLogin, (req,res) => {
 });
 
 app.get("/report/Delete/:reportID", ensureLogin, (req, res) => {
+    if(req.session.user.isAdmin === true){
     const reportID = req.params.reportID;
     ReportModel.deleteOne({_id: reportID})
         .then(()=>{
             res.redirect("/report");
         });
+    } else {
+        res.redirect("/Profile");
+    }
 })
 
 
@@ -382,6 +567,29 @@ app.get("/firstrunsetup", (req,res)=> {
             console.log("There was an error creating Yuri: " + err);
         } else {
             console.log("Yuri was created");
+        }
+    });
+    console.log("got here 2!");
+    res.redirect("/");
+})
+
+app.get("/secondrunsetup", (req,res)=> {
+    var Aziz = new UserModel({
+        username: 'aziz',
+        password: '123456',
+        firstName: 'Aziz',
+        lastName: 'Unknown',
+        email: 'unknown@gmail.com',
+        phone: '',
+        isAdmin: false
+    });
+    console.log("got here!");
+    Aziz.save((err)=> {
+        console.log("Error: " + err + ';');
+        if (err) {
+            console.log("There was an error creating Aziz: " + err);
+        } else {
+            console.log("Aziz was created");
         }
     });
     console.log("got here 2!");
